Update xterm, xterm-addon-fit and fix font problem
There was some problem with the font rendering, which seems to be because
it xterm calculates the font character widths before the font have been
downloaded. This commit adds xterm-webfont which waits with initializing
the xterm until all fonts have been loaded.

It's not the cleanest code ever, but good enough.