Port project to Vue (#12)
In order to be able to do some more extensive design, I've decided to rebuild stuga.org using Vue.
Vue was selected because for the learning experience, as I've never used it before. With this commit,
most of the previous functionality should be in place, and some preliminary navigation in place.