Add command to start game automatically, plus final Vue commit (#13)
This commit will start the game automatically by executing a `RunStatement` using the new `run` function in jomag/bajsic.
It will also add some functionality to preload the interpreter with some input data and option to convert all input to upper
case. Finally, it will autofocus the terminal.

This will also be the last Vue commit. After building some experience, I've come to the conclusion that I will instead do
a port to React. Some of the reasons are: better TypeScript support in React, the separation of JS and template code
is confusing at best, and most importantly: I have much more experience in React and will be able to work faster. This
Vue port was primarily done as a learning experience, and it was a success at that.