Port Stuga to React (#16)
So I wanted to do some more fancy stuff and decided that now was the time to add a framework.
This commit ports the website to React, after a detour testing out Vue. It also adds some webbish
modals for the background info and focus the terminal view on the game, starting it automatically.
By default, keyboard input is translated to uppercase as that's what the game expects. A floating
menu has been added to show the background info modal and toggle caps lock.