Refactor terminal components (#21)
This commit will restructure the terminal handling quite a bit. Previously there was three
main components involved:

- BasicXTerm - a class that created an XTermJS instance and Bajsic environment
- BasicTerminal - a React component that created a BasicXTerm class and assigned it to a div element
- App, which did little but render a BasicTerminal component

The separation of concern was not clear and the dependency between BasicXTerm and BasicTerminal
was quite fragile. It was not clear where to place new features.

With this commit, BasicXTerm has been dropped and replaced by an XTerm addon, BasicAddon, which
handles the interpreter environment and input/output to the assigned XTerm instance.

BasicTerminal has been renamed XTermWrapper, and is now a simple React wrapper around XTerm.
It is not tied to the Basic interpreter in any way. The implementation was inspired by
[xterm-for-react](https://www.npmjs.com/package/xterm-for-react).

Finally, App has some added responsibility to initialize the terminal and BasicAddon.

Also, fixes #20: terminal not changing size after resize.